Restructure hub commands
