+++ commands/hub/invite/index.js
